refs #13108 - reduce number of auth requests which gateway makes
The problem is that GW makes auth request every time when we want to
request something from the cloud. On the right_rackspace gem level everything
works fine, because we store auth data in instance variables, but it
doesn't work together with GW. We need to cache auth data somewhere.
As a solution we can store auth data in thread's local variables.
So we actually will not do extra auth request if we have already done it
in current thread.